# swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/users/profile":{"get":{"operationId":"UserController_profile","summary":"profile","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/permission":{"patch":{"operationId":"UserController_updateUserPermission","summary":"update permission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserPermissionDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/notes":{"get":{"operationId":"NoteController_notes","summary":"user notes","parameters":[{"required":false,"name":"limit","in":"query","schema":{"type":"number"}},{"required":false,"name":"page","in":"query","schema":{"type":"number"}},{"required":false,"name":"sort","in":"query","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedNoteResponse"}}}}},"tags":["notes"],"security":[{"bearer":[]}]},"post":{"operationId":"NoteController_create","summary":"Create Note","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNoteDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteResponse"}}}}},"tags":["notes"],"security":[{"bearer":[]}]}},"/api/notes/{id}":{"get":{"operationId":"NoteController_getById","summary":"Get Note By Id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteResponse"}}}}},"tags":["notes"],"security":[{"bearer":[]}]},"patch":{"operationId":"NoteController_updateById","summary":"Update Note By Id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNoteDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteResponse"}}}}},"tags":["notes"],"security":[{"bearer":[]}]},"delete":{"operationId":"NoteController_deleteById","summary":"Delete Note By Id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteResponse"}}}}},"tags":["notes"],"security":[{"bearer":[]}]}}},"info":{"title":"Notes","description":"# Welcome to Notes API 🎉\n\n","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000","description":"local"},{"url":"https://notes-api.gerrymie.site","description":"staging"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"User":{"type":"object","properties":{"uid":{"type":"string"},"name":{"type":"string"},"photoUrl":{"type":"string"},"email":{"type":"string"},"isAdmin":{"type":"boolean"}},"required":["uid","name","photoUrl","email","isAdmin"]},"UpdateUserPermissionDto":{"type":"object","properties":{"isAdmin":{"type":"boolean"},"email":{"type":"string"}},"required":["isAdmin","email"]},"NoteResponse":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"user":{"type":"object"},"_id":{"type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"}},"required":["title","content","user","_id","created_at","updated_at"]},"PaginatedNoteResponse":{"type":"object","properties":{"docs":{"type":"array","items":{"$ref":"#/components/schemas/NoteResponse"}},"hasNextPage":{"type":"boolean"},"hasPrevPage":{"type":"boolean"},"limit":{"type":"number"},"offset":{"type":"number"},"pagingCounter":{"type":"number"},"totalDocs":{"type":"number"},"totalPages":{"type":"number"}},"required":["docs","hasNextPage","hasPrevPage","limit","offset","pagingCounter","totalDocs","totalPages"]},"CreateNoteDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}},"required":["title","content"]},"UpdateNoteDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}}}}}}